Refactor package of ConfigCenter related classes
