+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/config/service/SchemaBasedPersistService.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.registry.config.service;
+package org.apache.shardingsphere.governance.core.config.service;
 
 /**
  * Schema based persist service.
