+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/config/service/impl/DataSourcePersistService.java
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.registry.config.service.impl;
+package org.apache.shardingsphere.governance.core.config.service.impl;
 
 import com.google.common.base.Strings;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.governance.core.registry.config.node.SchemaMetadataNode;
-import org.apache.shardingsphere.governance.core.registry.config.service.SchemaBasedPersistService;
+import org.apache.shardingsphere.governance.core.config.node.SchemaMetadataNode;
+import org.apache.shardingsphere.governance.core.config.service.SchemaBasedPersistService;
 import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
 import org.apache.shardingsphere.infra.config.datasource.DataSourceConfiguration;
 import org.apache.shardingsphere.infra.yaml.engine.YamlEngine;
